+++ src/geometries/Rect.cc
@@ -23,7 +23,6 @@
     along with GPUSPH.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include <cmath>
 #include <cstdio>
 #include <cstdlib>
 #include <iostream>
@@ -50,7 +49,7 @@ Rect::Rect(void)
 Rect::Rect(const Point& origin, const Vector& vx, const Vector& vy)
 {
 	if (fabs(vx*vy) > 1.e-8*vx.norm()*vy.norm()) {
-		//std::cout << "Trying to construct a rectangle with non perpendicular vectors\n";
+		//cout << "Trying to construct a rectangle with non perpendicular vectors\n";
 		//exit(1);
 	}
 
@@ -332,13 +331,13 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp1 = vp + m_vx/(nx*3) + 2*m_vy/(ny*3);
 				bpoints.push_back(bp1);
 
-				uint4 vertices1 = {vpoint_index, vpoint_index + 1, vpoints.size() + i + 1, 0};
+				uint4 vertices1 = {vpoint_index, vpoint_index + 1, uint(vpoints.size() + i + 1), 0};
 				vindexes.push_back(vertices1);
 
 				Point bp2 = vp + 2*m_vx/(nx*3) + m_vy/(ny*3);
 				bpoints.push_back(bp2);
 
-				uint4 vertices2 = {vpoint_index, vpoints.size() + i, vpoints.size() + i + 1, 0};
+				uint4 vertices2 = {vpoint_index, uint(vpoints.size() + i), uint(vpoints.size() + i + 1), 0};
 				vindexes.push_back(vertices2);
 
 				belems.push_back(belm);
@@ -359,13 +358,13 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp1 = vp + m_vx/(nx*3) + 2*m_vy/(ny*3);
 				bpoints.push_back(bp1);
 
-				uint4 vertices1 = {vpoint_index, vpoint_index + 1, vpoints.size() + i + 1, 0};
+				uint4 vertices1 = {vpoint_index, vpoint_index + 1, uint(vpoints.size() + i + 1), 0};
 				vindexes.push_back(vertices1);
 
 				Point bp2 = vp + 2*m_vx/(nx*3) + m_vy/(ny*3);
 				bpoints.push_back(bp2);
 
-				uint4 vertices2 = {vpoint_index, vpoints.size() + i, vpoints.size() + i + 1, 0};
+				uint4 vertices2 = {vpoint_index, uint(vpoints.size() + i), uint(vpoints.size() + i + 1), 0};
 				vindexes.push_back(vertices2);
 
 				belems.push_back(belm);
@@ -387,13 +386,13 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp1 = vp + m_vx/(nx*3) + 2*m_vy/(ny*3);
 				bpoints.push_back(bp1);
 
-				uint4 vertices1 = {vpoint_index, vpoint_index + 1, vpoints.size() + i + 1, 0};
+				uint4 vertices1 = {vpoint_index, vpoint_index + 1, uint(vpoints.size() + i + 1), 0};
 				vindexes.push_back(vertices1);
 
 				Point bp2 = vp + 2*m_vx/(nx*3) + m_vy/(ny*3);
 				bpoints.push_back(bp2);
 
-				uint4 vertices2 = {vpoint_index, vpoints.size() + i, vpoints.size() + i + 1, 0};
+				uint4 vertices2 = {vpoint_index, uint(vpoints.size() + i), uint(vpoints.size() + i + 1), 0};
 				vindexes.push_back(vertices2);
 			}
 
@@ -406,13 +405,13 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp1 = vp - 2*m_vx/(nx*3) + 2*m_vy/(ny*3);
 				bpoints.push_back(bp1);
 
-				uint4 vertices1 = {vpoints.size() + (endy-starty+1)*(endx-startx) + i, vpoints.size() + (endy-starty+1)*(endx-startx)+ i + 1, vpoint_index + 1, 0};
+				uint4 vertices1 = {uint(vpoints.size() + (endy-starty+1)*(endx-startx) + i), uint(vpoints.size() + (endy-starty+1)*(endx-startx)+ i + 1), vpoint_index + 1, 0};
 				vindexes.push_back(vertices1);
 
 				Point bp2 = vp - m_vx/(nx*3) + m_vy/(ny*3);
 				bpoints.push_back(bp2);
 
-				uint4 vertices2 = {vpoints.size() + (endy-starty+1)*(endx-startx) + i, vpoint_index, vpoint_index + 1, 0};
+				uint4 vertices2 = {uint(vpoints.size() + (endy-starty+1)*(endx-startx) + i), vpoint_index, vpoint_index + 1, 0};
 				vindexes.push_back(vertices2);
 			}
 
@@ -437,13 +436,13 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp1 = vp + 2*m_vx/(nx*3) + m_vy/(ny*3);
 				bpoints.push_back(bp1);
 
-				uint4 vertices1 = {vpoint1_index, vpoints.size() + (endy-starty) - i, vpoints.size() + (endy-starty) - i + 1, 0};
+				uint4 vertices1 = {vpoint1_index, uint(vpoints.size() + (endy-starty) - i), uint(vpoints.size() + (endy-starty) - i + 1), 0};
 				vindexes.push_back(vertices1);
 
 				Point bp2 = vp + m_vx/(nx*3) + 2*m_vy/(ny*3);
 				bpoints.push_back(bp2);
 
-				uint4 vertices2 = {vpoint0_index, vpoint1_index, vpoints.size() + (endy-starty) - i + 1, 0};
+				uint4 vertices2 = {vpoint0_index, vpoint1_index, uint(vpoints.size() + (endy-starty) - i + 1), 0};
 				vindexes.push_back(vertices2);
 			}
 
@@ -457,7 +456,7 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp1 = vp + 2*m_vx/(nx*3) - 2*m_vy/(ny*3);
 				bpoints.push_back(bp1);
 
-				uint4 vertices1 = {vpoint0_index, vpoints.size()-1 + (endy-starty+1)*(endx-startx-i+1), vpoints.size()-1 + (endy-starty+1)*(endx-startx-i+2), 0};
+				uint4 vertices1 = {vpoint0_index, uint(vpoints.size()-1 + (endy-starty+1)*(endx-startx-i+1)), uint(vpoints.size()-1 + (endy-starty+1)*(endx-startx-i+2)), 0};
 				if(i == predef_vparts-1)
 					vertices1.y = edgeparts[3][1][0];
 				vindexes.push_back(vertices1);
@@ -465,7 +464,7 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp2 = vp + m_vx/(nx*3) - 1*m_vy/(ny*3);
 				bpoints.push_back(bp2);
 
-				uint4 vertices2 = {vpoint0_index, vpoints.size()-1 + (endy-starty+1)*(endx-startx-i+1), vpoint1_index, 0};
+				uint4 vertices2 = {vpoint0_index, uint(vpoints.size()-1 + (endy-starty+1)*(endx-startx-i+1)), vpoint1_index, 0};
 				if(i == predef_vparts-1)
 					vertices2.y = edgeparts[3][1][0];
 				vindexes.push_back(vertices2);
@@ -481,13 +480,13 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp1 = vp - m_vx/(nx*3) - 2*m_vy/(ny*3);
 				bpoints.push_back(bp1);
 
-				uint4 vertices1 = {vpoint1_index, vpoint0_index, vpoints.size()-1 + (endy-starty+1)*(endx-startx) + i, 0};
+				uint4 vertices1 = {vpoint1_index, vpoint0_index, uint(vpoints.size()-1 + (endy-starty+1)*(endx-startx) + i), 0};
 				vindexes.push_back(vertices1);
 
 				Point bp2 = vp - 2*m_vx/(nx*3) - m_vy/(ny*3);
 				bpoints.push_back(bp2);
 
-				uint4 vertices2 = {vpoint1_index,vpoints.size()-1 + (endy-starty+1)*(endx-startx) + i,vpoints.size()-1 + (endy-starty+1)*(endx-startx) + i + 1, 0};
+				uint4 vertices2 = {vpoint1_index, uint(vpoints.size()-1 + (endy-starty+1)*(endx-startx) + i), uint(vpoints.size()-1 + (endy-starty+1)*(endx-startx) + i + 1), 0};
 				if(i == predef_vparts-1)
 					vertices2.z = edgeparts[2][1][0];
 				vindexes.push_back(vertices2);
@@ -503,7 +502,7 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp1 = vp - m_vx/(nx*3) + m_vy/(ny*3);
 				bpoints.push_back(bp1);
 
-				uint4 vertices1 = {vpoint0_index, vpoint1_index, vpoints.size() + (endy-starty+1)*(i-1), 0};
+				uint4 vertices1 = {vpoint0_index, vpoint1_index, uint(vpoints.size() + (endy-starty+1)*(i-1)), 0};
 				if(i == predef_vparts-1)
 					vertices1.z = edgeparts[1][1][0];
 				vindexes.push_back(vertices1);
@@ -511,7 +510,7 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp2 = vp - 2*m_vx/(nx*3) + 2*m_vy/(ny*3);
 				bpoints.push_back(bp2);
 
-				uint4 vertices2 = {vpoint0_index, vpoints.size() + (endy-starty+1)*(i-2), vpoints.size() + (endy-starty+1)*(i-1), 0};
+				uint4 vertices2 = {vpoint0_index, uint(vpoints.size() + (endy-starty+1)*(i-2)), uint(vpoints.size() + (endy-starty+1)*(i-1)), 0};
 				if(i == 1)
 					vertices2.y = edgeparts[3][1][endx-startx];
 				if(i == predef_vparts-1)
@@ -537,7 +536,7 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 	for (int i = startx; i <= endx; i++)
 	for (int j = starty; j <= endy; j++) {
 			Point vp = m_origin + i*m_vx/nx + j*m_vy/ny;
-			int nvertex = vpoints.size();
+			uint nvertex = vpoints.size();
 
 			//Save vertex particles located at the edges of planes
 			if (i == 0) {
@@ -563,13 +562,13 @@ Rect::Fill(PointVect& bpoints, PointVect& belems, PointVect& vpoints, std::vecto
 				Point bp1 = vp + m_vx/(nx*3) + 2*m_vy/(ny*3);
 				bpoints.push_back(bp1);
 
-				uint4 vertices1 = {nvertex, nvertex + 1, nvertex + (endy-starty) + 2, 0};
+				uint4 vertices1 = {nvertex, nvertex + 1, uint(nvertex + (endy-starty) + 2), 0};
 				vindexes.push_back(vertices1);
 
 				Point bp2 = vp + 2*m_vx/(nx*3) + m_vy/(ny*3);
 				bpoints.push_back(bp2);
 
-				uint4 vertices2 = {nvertex, nvertex + (endy-starty) + 1, nvertex + (endy-starty) + 2, 0};
+				uint4 vertices2 = {nvertex, uint(nvertex + (endy-starty) + 1), uint(nvertex + (endy-starty) + 2), 0};
 				vindexes.push_back(vertices2);
 
 				belems.push_back(belm);
